refactor: restructure project for web extension - Move TUI entry point to cmd/tui/main.go - Move core business logic to pkg/core/ - Move TUI components to internal/tui/ - Update all import paths and build instructions - Move integration tests to cmd/tui/main_test.go - Update README with new project structure and build commands - All tests pass and TUI functionality verified
