+++ internal/tui/ui.go
@@ -11,9 +11,9 @@ import (
 	"github.com/charmbracelet/bubbles/textinput"
 	tea "github.com/charmbracelet/bubbletea"
 	"github.com/charmbracelet/lipgloss"
-	"github.com/lauren/nannytracker/internal/maps"
-	"github.com/lauren/nannytracker/internal/model"
-	"github.com/lauren/nannytracker/internal/storage"
+	"github.com/laurendc/nannytracker/pkg/core/maps"
+	model "github.com/laurendc/nannytracker/pkg/core"
+	"github.com/laurendc/nannytracker/pkg/core/storage"
 )
 
 // Model represents the UI state
