+++ pkg/core/maps/client_test.go
