+++ pkg/core/model_test.go
