+++ pkg/core/trip.go
