+++ pkg/core/trip_test.go
